Extract parser logic from _error.js
Extracted the logic which parsed the html response provided by drupal to
htmlResponseParser.js
