+++ app/utils/markup.js
@@ -1,0 +1,3 @@
+export function createHtmlMarkup(html) {
+  return {__html: html};
+}
